H5 checkpoint: seeded reconvert at go1.24.13 + H5c (item 11) -- NOT the H5 gate: stdlib slnx lists 14 deleted packages (MSB3202), 5 hand-owns orphaned; see mailbox 6d5696dcd

+++ src/core/encoding/binary/varint.cs
@@ -73,7 +73,7 @@ public static (uint64, nint) Uvarint(slice<byte> buf) {
             return (0, -(i + 1)); // overflow
         }
         if (b < 0x80) {
-            if (i == MaxVarintLen64 - 1 && b > 1) {
+            if (i == (nint)(MaxVarintLen64 - 1) && b > 1) {
                 return (0, -(i + 1)); // overflow
             }
             return ((uint64)(x | ((uint64)b).Lsh(s)), i + 1);
@@ -137,7 +137,7 @@ public static (uint64, error) ReadUvarint(io.ByteReader r) {
             return (x, err);
         }
         if (b < 0x80) {
-            if (i == MaxVarintLen64 - 1 && b > 1) {
+            if (i == (nint)(MaxVarintLen64 - 1) && b > 1) {
                 return (x, errOverflow);
             }
             return ((uint64)(x | ((uint64)b).Lsh(s)), default!);
