+++ src/core/math/floor.cs
@@ -96,11 +96,11 @@ public static float64 Round(float64 x) {
     if (e < bias){
         // Round abs(x) < 1 including denormals.
         bits &= (uint64)(signMask); // +-0
-        if (e == bias - 1) {
+        if (e == (nuint)(bias - 1)) {
             bits |= (uint64)(uvone); // +-1
         }
     } else 
-    if (e < bias + shift) {
+    if (e < (nuint)(bias + shift)) {
         // Round any abs(x) >= 1 containing a fractional component [0,1).
         //
         // Numbers with larger exponents are returned unchanged since they
@@ -143,7 +143,7 @@ public static float64 RoundToEven(float64 x) {
         bits += (halfMinusULP + (uint64)((bits.Rsh(((nuint)shift - e))) & 1)).Rsh(e);
         bits &= unchecked((uint64)~(uint64)(((uint64)fracMask).Rsh(e)));
     } else 
-    if (e == bias - 1 && (uint64)(bits & (uint64)fracMask) != 0){
+    if (e == (nuint)(bias - 1) && (uint64)(bits & (uint64)fracMask) != 0){
         // Round 0.5 < abs(x) < 1.
         bits = (uint64)((uint64)(bits & (uint64)signMask) | (uint64)uvone); // +-1
     } else {
